requirements.txt
#You should install libs for your environment, 
#then you can import in inside te container.
requests
Pillow
